patterns/langgraph-ag-ui-agent/requirements.txt
# LangGraph AG-UI agent dependencies with pinned versions
fastapi==0.115.12
uvicorn==0.32.1
ag-ui-langgraph==0.0.25
langgraph==1.0.8
langchain-aws==1.0.0
copilotkit==0.1.77
langgraph-checkpoint-aws==1.0.1
